+++ feat/ewgf/src/commonMain/kotlin/io/github/sophon/domain/EwgfError.kt
@@ -4,8 +4,8 @@ import io.github.sophon.core.domain.Error
 
 sealed class EwgfError(vararg val errors: String): Error {
     class DatabaseError(error: String): EwgfError(error)
-    class PlayerNotFound(query: String): EwgfError(query)
-    class PlayerAlreadyRegistered(query: String): EwgfError(query)
+    class PlayerNotRegistered(id: String): EwgfError(id)
+    class PlayerNotFound(id: String): EwgfError(id)
     class LogicError(error: String): EwgfError(error)
 
     override fun toString(): String =